Rename to PolicyAnalzerClient
