+++ sources/src/main/java/com/google/solutions/jitaccess/web/RuntimeEnvironment.java
@@ -182,7 +182,7 @@ public class RuntimeEnvironment {
             null,
             Stream.of(
                 ResourceManagerClient.OAUTH_SCOPE,
-                AssetInventoryClient.OAUTH_SCOPE,
+                PolicyAnalyzerClient.OAUTH_SCOPE,
                 IamCredentialsClient.OAUTH_SCOPE,
                 SecretManagerClient.OAUTH_SCOPE,
                 DirectoryGroupsClient.OAUTH_SCOPE)
